change ["-1", "0", "1"] value ranges to ["0", "1", "2"] for consistency
Some controls (notably rocker switches, but also some three-position toggle
switches) reported their state as "-1", "0" or "1". Those have been changed
to "0", "1" and "2", where "1" is the neutral position.

This change makes it consistent with other three-position controls and also
ensures that each state takes only one character, which makes it slightly
more convenient to deal with on an 8-bit microcontroller.