change license to a slightly modified Simple Public License 2.0
With this commit, I am changing the license from MIT to a modified
SimPL-2.0. I want a "strong copyleft" license, but I also want a
license that gets out of my way, which is the reason I initially used
MIT.

The GPL does not fulfill the "get out of my way" requirement. While it is
widely used and comparatively well understood, it does not change the fact
that it is littered with undecipherable legalese. It also spends a lot of
words talking about binary distribution, which is irrelevant to a project
that does not use any compiled languages.

The GPL also requires you to prominently note the date of any changes
you make. While I agree it is a good idea to do so, IMO the place for a
change log is not the work itself (or *shudder* the individual source
files, as required by GPLv2), but a revision control system.

To avoid any confusion about that part, I modified the SimPL and removed
the requirement to note the date of changes entirely.