Revert "Add DCSBIOS_ADD_EXPORT_TARGET and DCSBIOS_REMOVE_EXPORT_TARGET commands"
This reverts commit ae772c9f700f93d59493bec8724f11695ed4d612.

While trying to implement this on the client side it became
apparent that this changeset was not fully thought through.

There is no way to reliably stop exporting to the additional
target IP/Port once that client has been closed. Doing so
would require implementing a keep-alive mechanism.

We have TCP support now, which mostly solves the same problem
in a simpler way (with the disadvantage of a very small
performance penalty).

Additional UDP targets can still be used by statically configuring
them in BIOSConfig.lua.

Closes #6.