rev incl more aircraft
added more of the supported aircraft to the list, to include all official modules, plus several of the higher profile mods. Note, button presses do not seem to be recognized for FC3 aircraft anymore. Not sure if I broke something with this mod, or if it was some other change in the most recent release.