#50 Right-align IFEI fuel text
This could be a breaking change depending on how users are parsing this data. However, it's definitely more correct to right-align them and will result in significantly fewer headaches for new users.

This also fixes a bug in `padLeft()` which could result in `nil` being actually displayed as the text `"nil"`