#151 Fix simshaker incompatibility
Removes attempt to stringify a global variable which is never set (but can be set by other plugins). SimShaker had a table stored in a global variable `msg` for some reason, and when `setValue()` was called with a`nil` argument (which happens sometimes in the A-10, though that may be a separate issue) it attempted to format the table as a string, which is not supported.