Require tests to pass before running nightly release (#260)
Previously the release would run even if tests would fail. This could be problematic as the tests could fail due to errors that would prevent dcs-bios from working, and as such we should make sure not to release a new nightly build in this case as it would be non-functional.