Don't export MetadataStart/End with any modules
MetadataStart and End are explicitly run before and after the main module exports, respectively. This is part of what bios guarantees - that MetadataStart will _always_ come first, and MetadataEnd will _always_ come last.

As these exports are explicitly run in Protocol.lua for that reason, we don't want to export them with the standard module export code.